+++ src/SlackConnector.Tests.Unit/SlackConnectionTests/Connect/TeamAndSelfTests.cs
@@ -1,14 +1,11 @@
 ﻿using Moq;
 using NUnit.Framework;
 using Should;
-using SlackConnector.Connections;
 using SlackConnector.Connections.Handshaking;
 using SlackConnector.Connections.Models;
-using SlackConnector.Connections.Sockets;
-using SlackConnector.Tests.Unit.SlackConnectorTests.Setups;
-using SpecsFor;
+using SlackConnector.Tests.Unit.SlackConnectionTests.Setups;
 
-namespace SlackConnector.Tests.Unit.SlackConnectorTests.Connect
+namespace SlackConnector.Tests.Unit.SlackConnectionTests.Connect
 {
     public static class TeamAndSelfTests
     {
