+++ src/SlackConnector.Tests.Unit/SlackConnectionTests/Connect/UsersTests.cs
@@ -3,9 +3,9 @@ using NUnit.Framework;
 using Should;
 using SlackConnector.Connections.Handshaking;
 using SlackConnector.Connections.Models;
-using SlackConnector.Tests.Unit.SlackConnectorTests.Setups;
+using SlackConnector.Tests.Unit.SlackConnectionTests.Setups;
 
-namespace SlackConnector.Tests.Unit.SlackConnectorTests.Connect
+namespace SlackConnector.Tests.Unit.SlackConnectionTests.Connect
 {
     public static class UsersTests
     {
