+++ src/SlackConnector.Tests.Unit/SlackConnector.Tests.Unit.csproj
@@ -93,16 +93,16 @@
     <Compile Include="SlackConnectorFactoryTests\TeamAndSelfTests.cs" />
     <Compile Include="SlackConnectorFactoryTests\UsersTests.cs" />
     <Compile Include="SlackConnectorFactoryTests\WebSocketTests.cs" />
-    <Compile Include="SlackConnectorTests\Connect\ConnectedStatusTests.cs" />
-    <Compile Include="SlackConnectorTests\InboundMessageTests.cs" />
-    <Compile Include="SlackConnectorTests\Connect\MultipleConnections.cs" />
-    <Compile Include="SlackConnectorTests\Connect\HubsTests.cs" />
-    <Compile Include="SlackConnectorTests\JoinDirectMessageChannelTests.cs" />
-    <Compile Include="SlackConnectorTests\SayTests.cs" />
-    <Compile Include="SlackConnectorTests\Setups\SlackConnectorIsSetup.cs" />
-    <Compile Include="SlackConnectorTests\Connect\UsersTests.cs" />
-    <Compile Include="SlackConnectorTests\Connect\TeamAndSelfTests.cs" />
-    <Compile Include="SlackConnectorTests\Connect\WebSocketTests.cs" />
+    <Compile Include="SlackConnectionTests\Connect\ConnectedStatusTests.cs" />
+    <Compile Include="SlackConnectionTests\InboundMessageTests.cs" />
+    <Compile Include="SlackConnectionTests\Connect\MultipleConnections.cs" />
+    <Compile Include="SlackConnectionTests\Connect\HubsTests.cs" />
+    <Compile Include="SlackConnectionTests\JoinDirectMessageChannelTests.cs" />
+    <Compile Include="SlackConnectionTests\SayTests.cs" />
+    <Compile Include="SlackConnectionTests\Setups\SlackConnectorIsSetup.cs" />
+    <Compile Include="SlackConnectionTests\Connect\UsersTests.cs" />
+    <Compile Include="SlackConnectionTests\Connect\TeamAndSelfTests.cs" />
+    <Compile Include="SlackConnectionTests\Connect\WebSocketTests.cs" />
     <Compile Include="Connections\Sockets\Messages\MessageInterpreterTests.cs" />
     <Compile Include="Stubs\RestClientStub.cs" />
     <Compile Include="Stubs\WebSocketClientStub.cs" />
