+++ src/SlackConnector.Tests.Unit/SlackConnectorFactoryTests/ConnectedStatusTests.cs
@@ -6,7 +6,7 @@ using Should;
 using SlackConnector.Connections;
 using SlackConnector.Connections.Handshaking;
 using SlackConnector.Exceptions;
-using SlackConnector.Tests.Unit.SlackConnectorTests.Setups;
+using SlackConnector.Tests.Unit.SlackConnectionTests.Setups;
 
 namespace SlackConnector.Tests.Unit.SlackConnectorFactoryTests
 {
