+++ src/SlackConnector.Tests.Unit/SlackConnectorFactoryTests/MultipleConnections.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using NUnit.Framework;
 using SlackConnector.Exceptions;
-using SlackConnector.Tests.Unit.SlackConnectorTests.Setups;
+using SlackConnector.Tests.Unit.SlackConnectionTests.Setups;
 
 namespace SlackConnector.Tests.Unit.SlackConnectorFactoryTests
 {
