+++ src/SlackConnector.Tests.Unit/SlackConnectorFactoryTests/TeamAndSelfTests.cs
@@ -3,7 +3,7 @@ using NUnit.Framework;
 using Should;
 using SlackConnector.Connections.Handshaking;
 using SlackConnector.Connections.Models;
-using SlackConnector.Tests.Unit.SlackConnectorTests.Setups;
+using SlackConnector.Tests.Unit.SlackConnectionTests.Setups;
 
 namespace SlackConnector.Tests.Unit.SlackConnectorFactoryTests
 {
